Centralized tests setup into XCTestCase extension
