+++ CommercetoolsTests/AuthManagerTests.swift
@@ -18,21 +18,6 @@ class AuthManagerTests: XCTestCase {
         
         super.tearDown()
     }
-
-    private func cleanPersistedTokens() {
-        let tokenStore = AuthManager.sharedInstance.tokenStore
-        tokenStore.accessToken = nil
-        tokenStore.refreshToken = nil
-        tokenStore.tokenValidDate = nil
-    }
-    
-    private func setupTestConfiguration() {
-        let testBundle = NSBundle(forClass: AuthManagerTests.self)
-        if let path = testBundle.pathForResource("CommercetoolsTestConfig", ofType: "plist"),
-            config = NSDictionary(contentsOfFile: path) {
-            Commercetools.config = Config(config: config)
-        }
-    }
     
     func testUserLogin() {
         setupTestConfiguration()
