+++ CommercetoolsTests/ByIdEndpointTests.swift
@@ -26,22 +26,7 @@ class ByIdEndpointTests: XCTestCase {
         cleanPersistedTokens()
         super.tearDown()
     }
-    
-    private func setupTestConfiguration() {
-        let testBundle = NSBundle(forClass: AuthManagerTests.self)
-        if let path = testBundle.pathForResource("CommercetoolsTestConfig", ofType: "plist"),
-            config = NSDictionary(contentsOfFile: path) {
-            Commercetools.config = Config(config: config)
-        }
-    }
-    
-    private func cleanPersistedTokens() {
-        let tokenStore = AuthManager.sharedInstance.tokenStore
-        tokenStore.accessToken = nil
-        tokenStore.refreshToken = nil
-        tokenStore.tokenValidDate = nil
-    }
-    
+
     func testByIdEndpoint() {
         
         let byIdExpectation = expectationWithDescription("byId expectation")
