+++ CommercetoolsTests/CreateEndpointTests.swift
@@ -23,21 +23,6 @@ class CreateEndpointTests: XCTestCase {
         super.tearDown()
     }
 
-    private func setupTestConfiguration() {
-        let testBundle = NSBundle(forClass: AuthManagerTests.self)
-        if let path = testBundle.pathForResource("CommercetoolsTestConfig", ofType: "plist"),
-        config = NSDictionary(contentsOfFile: path) {
-            Commercetools.config = Config(config: config)
-        }
-    }
-
-    private func cleanPersistedTokens() {
-        let tokenStore = AuthManager.sharedInstance.tokenStore
-        tokenStore.accessToken = nil
-        tokenStore.refreshToken = nil
-        tokenStore.tokenValidDate = nil
-    }
-
     func testCreateEndpoint() {
 
         let createExpectation = expectationWithDescription("create expectation")
